This is the full package
here is a full upload and working model of the overlay mod, please read the readme to get the full rundown on how to use. I also have a .md wrote to explain all the settings in config.py because some may not know exactly what each do and it just makes this whole mod more cohesive.