Update README.md
Added deps to the readme, explained what they are and what they do and why we need them. I also put the Bash code in to the readme just in case, I also wrote up a python script for it all were it will download all the deps and python 3.13. There will be more changes soon.