Update README.md
Added more info and structured the way I had the .md and structure setup. I did not like how confusing it was before, now there is a new -Setup Info section, section is to breakdown what is needed and what info you need to know before doing what is needed to use this mod.