Preflight checks update
This was additions to the main.py adding a preflight check for the deps in Python, also it runs alongside required_downloads.py, to check that you have the files that I require of you, plus it will also ask you if it's okay to download the deps for you. I added this because so many times over the years of playing others mods, I have ran into issues were I have no idea what i'm missing. So I copied steam and now it asks u if downloading pip deps for Python is okay, im working hard to take the guesswork out of using a mod, it's always a huge come off when you have to struggle to get something to work. I also tried to write a script that tests the deps and checks, its said it passed, but felt very elementary. Tell me if you want the files ill send them to you.