.png files that were missing
Here are the PNG files that are from beta version of the game, these cells I think are not in the non beta version of the game, but idk for sure. But here they are, not perfect by any means, but they are center but ill need to do more editing on them so keep eye out for an update in the next few days to make them better and 90x90. 
ADDED↓
Cell_wandering.png - Wandering_neuron
Cell_ink.png - Ink_cell  
Cell_impact.png - Impact_cell 
Cell_magnetic.png - Electromagnetic_cell