Crop and size change
Noticed, that even at 120 when you build was long portrait up and down, the top and bottoms of the build would get cut off. There is more I need to do but this is a simple hotfix. I also increased the Hex_size to 25 looked a little bit better like this!